Revises report to significantly reduce text, refactor files
